+++ lib/constants/constants.dart
@@ -1,1 +1,3 @@
 // export semua file dari folder constant simpan disini
+
+export 'dimens.dart';
