+++ lib/constants/dimens.dart
@@ -1,0 +1,5 @@
+import 'package:get/get.dart';
+
+final double paddingTop = Get.context.mediaQueryPadding.top;
+final double widthApp = Get.context.width;
+final double heightApp = Get.context.height - paddingTop;
