+++ lib/ui/components/components.dart
@@ -1,0 +1,4 @@
+// export semua file dari folder ui simpan disini
+
+export 'xetia_bottom_bar.dart';
+export 'xetia_tab_bar.dart';
