+++ lib/ui/ui.dart
@@ -1,1 +1,7 @@
+// export semua file dari folder ui simpan disini
+
 export 'home_ui.dart';
+export 'shop_ui.dart';
+export 'favorite_ui.dart';
+export 'cart_ui.dart';
+export 'setting_ui.dart';
